Updated sidenav to navigate to Device page
Added additional colours to the dashboard as well to match the green theme. The sidenav only includes a link to the new page to be able to view the changes but there is no way yet to navigate to the page via a component yet. Will need backend functionality populated components for that.